requirements.txt
numpy>=1.24
scipy>=1.10
pyyaml>=6.0
tqdm>=4.66
torch>=2.1

